cleaning up airflow pvs after shutdown

+++ scripts/airflow.sh
@@ -114,7 +114,7 @@ start() {
 
     create_namespace "$NAMESPACE"
     create_secrets "$NAMESPACE"
-\
+
     if ! build_and_load_image "$DIR" "$IMAGE_REPO" "$IMAGE_TAG" ; then
         echo "Failed to load image to local registry"
         exit 1
@@ -145,14 +145,32 @@ start() {
 
 # shutdown function
 shutdown() {
-    kubectl delete namespace "$NAMESPACE"
-
     local app="airflow"
     local env_file="$STORAGE_DIR/.env.$app"
 
+
+    # check if namespace exists, and if it does, delete it
+    if kubectl get namespace "$NAMESPACE" &> /dev/null; then
+        kubectl delete namespace "$NAMESPACE"
+    fi
+
     # Check if .env file exists 
     shutdown_docker_compose_stack "$app" "$env_file" "$DELETE_DATA" "$DOCKER_COMPOSE_FILE"
-}
+
+    # Delete persistent volumes
+    # Get all persistent volumes with label app=airflow
+    pvs=$(kubectl get pv -l app=airflow -o jsonpath="{.items[*].metadata.name}")
+
+    echo "Deleting persistent volumes"
+    echo "$pvs"
+
+    # Iterate over the list of persistent volumes
+    for pv in $pvs
+    do
+        # Delete each persistent volume
+        kubectl delete pv $pv
+    done
+    }
 
 init(){
     # copy .env file if it doesn't exist
